chore: fixed linting, moved helper function
