+++ admin/src/helpers/blob.js
@@ -1,0 +1,9 @@
+export function b64toBlob(dataURI, type) {
+    const byteString = atob(dataURI);
+    const ab = new ArrayBuffer(byteString.length);
+    const ia = new Uint8Array(ab);
+    for (let i = 0; i < byteString.length; i++) {
+        ia[i] = byteString.charCodeAt(i);
+    }
+    return new Blob([ab], { type });
+}
